+++ src/Cli/func/Telemetry/Sha256Hasher.cs
@@ -1,4 +1,4 @@
-// Copyright (c) .NET Foundation. All rights reserved.
+﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System.Security.Cryptography;
@@ -9,7 +9,7 @@ namespace Azure.Functions.Cli.Telemetry
     internal static class Sha256Hasher
     {
         /// <summary>
-        /// The hashed mac address needs to be the same hashed value as produced by the other distinct sources given the same input. (e.g. VsCode)
+        /// The hashed mac address needs to be the same hashed value as produced by the other distinct sources given the same input. (e.g. VsCode).
         /// </summary>
         public static string Hash(string text)
         {