+++ test/Cli/Func.UnitTests/ActionsTests/StartHostActionTests.cs
@@ -4,14 +4,21 @@
 using System.IO.Abstractions;
 using System.Runtime.InteropServices;
 using System.Text;
+
 using Azure.Functions.Cli.Actions.HostActions;
 using Azure.Functions.Cli.Common;
 using Azure.Functions.Cli.Helpers;
 using Azure.Functions.Cli.Interfaces;
+
 using Colors.Net;
+
 using FluentAssertions;
+using FluentAssertions.Execution;
+
 using Moq;
+
 using NSubstitute;
+
 using Xunit;
 
 namespace Azure.Functions.Cli.UnitTests.ActionsTests
@@ -259,6 +266,42 @@ namespace Azure.Functions.Cli.UnitTests.ActionsTests
             }
 
             Assert.False(expectException, "Expected validation failure.");
+        }
+
+        [InlineData(WorkerRuntime.DotnetIsolated)]
+        [InlineData(WorkerRuntime.Python)]
+        [InlineData(WorkerRuntime.Java)]
+        [InlineData(WorkerRuntime.Node)]
+        [Theory]
+        public async Task AzureFunctionsEnvironment_EnvironmentVariable_SetByUser_DoesNotThrow(WorkerRuntime currentRuntime)
+        {
+            Environment.SetEnvironmentVariable(Common.Constants.FunctionsWorkerRuntime, WorkerRuntimeLanguageHelper.GetRuntimeMoniker(currentRuntime));
+            GlobalCoreToolsSettings.SetWorkerRuntime(currentRuntime);
+
+            Mock<IProcessManager> processManager = new();
+            Mock<ISecretsManager> secretsManager = new();
+
+            var settings = new Dictionary<string, string> { ["AZURE_FUNCTIONS_ENVIRONMENT"] = "MyEnvironment" };
+            secretsManager.Setup(s => s.GetSecrets()).Returns(() => settings);
+
+            GlobalCoreToolsSettings.Init(secretsManager.Object, []);
+
+            var startHostAction = new StartHostAction(secretsManager.Object, processManager.Object)
+            {
+                HostRuntime = "default",
+                VerboseLogging = true,
+            };
+
+            try
+            {
+                await startHostAction.RunAsync().WaitAsync(TimeSpan.FromSeconds(1));
+                throw new AssertionFailedException("Should've canceled via timeout");
+            }
+            catch (TimeoutException)
+            {
+            }
+
+            Assert.Equal("MyEnvironment", settings["AZURE_FUNCTIONS_ENVIRONMENT"]);
         }
 
         public void Dispose()