+++ release_notes.md
@@ -8,3 +8,4 @@
 #### Changes
 
 - Fix dotnet templates installation (#5358)
+- Fix exception when `AZURE_FUNCTIONS_ENVIRONMENT` variable is set by user (#2465)